feat(be): 프로젝트 권한 예외처리 추가

+++ backend/src/main/java/com/ourhour/domain/project/service/MilestoneService.java
@@ -3,6 +3,7 @@ package com.ourhour.domain.project.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.ourhour.domain.org.enums.Role;
 import com.ourhour.domain.project.dto.MilestoneReqDTO;
 import com.ourhour.domain.project.entity.MilestoneEntity;
 import com.ourhour.domain.project.entity.ProjectEntity;
@@ -11,6 +12,7 @@ import com.ourhour.domain.project.repository.MilestoneRepository;
 import com.ourhour.domain.project.repository.ProjectRepository;
 import com.ourhour.global.common.dto.ApiResponse;
 import com.ourhour.global.exception.BusinessException;
+import com.ourhour.global.jwt.dto.Claims;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,6 +25,7 @@ public class MilestoneService {
     private final MilestoneRepository milestoneRepository;
     private final ProjectRepository projectRepository;
     private final MilestoneMapper milestoneMapper;
+    private final ProjectParticipantService projectParticipantService;
 
     // 마일스톤 등록
     @Transactional
@@ -66,10 +69,29 @@ public class MilestoneService {
 
     // 마일스톤 삭제
     @Transactional
-    public ApiResponse<Void> deleteMilestone(Long milestoneId) {
+    public ApiResponse<Void> deleteMilestone(Long milestoneId, Claims claims) {
 
         MilestoneEntity milestoneEntity = milestoneRepository.findById(milestoneId)
                 .orElseThrow(() -> BusinessException.badRequest("존재하지 않는 마일스톤 ID입니다."));
+
+        Long orgId = milestoneEntity.getProjectEntity().getOrgEntity().getOrgId();
+        Long projectId = milestoneEntity.getProjectEntity().getProjectId();
+
+        Long memberId = claims.getOrgAuthorityList().stream()       
+                                .filter(auth -> auth.getOrgId().equals(orgId))
+                                .map(auth -> auth.getMemberId())
+                                .findFirst()
+        .orElseThrow(() -> BusinessException.forbidden("해당 회사의 멤버가 아닙니다."));
+
+        boolean isParticipant = projectParticipantService.isProjectParticipant(projectId, memberId);
+
+        boolean isAdminOrRootAdmin = claims.getOrgAuthorityList().stream()
+                                .filter(auth -> auth.getOrgId().equals(orgId))
+                                .anyMatch(auth -> auth.getRole().equals(Role.ADMIN) || auth.getRole().equals(Role.ROOT_ADMIN));
+
+        if (!(isParticipant || isAdminOrRootAdmin)) {
+                throw BusinessException.forbidden("프로젝트 참여자이거나 ADMIN 이상 권한이 있어야 합니다.");
+        }
 
         milestoneRepository.delete(milestoneEntity);
